added simulator tests. further restructuring.
