+++ simulator/tests/CMakeLists.txt
@@ -1,6 +1,11 @@
 project(test_simulator)
 
-add_executable(${PROJECT_NAME} test_ap_estimation.cpp test_item_optimizer.cpp)
+add_executable(${PROJECT_NAME}
+        test_ap_estimation.cpp
+        test_item_optimizer.cpp
+        test_use_effects.cpp
+        test_simulator.cpp
+        )
 
 target_link_libraries(${PROJECT_NAME} gtest_main wow_library simulator)
