+++ statistics/CMakeLists.txt
@@ -4,6 +4,7 @@ project(statistics)
 add_library(
         ${PROJECT_NAME}
         source/Statistics.cpp
+        source/Distribution.cpp
 )
 
 target_include_directories(${PROJECT_NAME} PUBLIC include ${CMAKE_CURRENT_SOURCE_DIR})
