+++ statistics/include/Distribution.h
@@ -1,0 +1,12 @@
+//
+// Created by Oliver on 2021-04-11.
+//
+
+#ifndef WOW_SIMULATOR_DISTRIBUTION_H
+#define WOW_SIMULATOR_DISTRIBUTION_H
+
+class Distribution
+{
+};
+
+#endif // WOW_SIMULATOR_DISTRIBUTION_H
