+++ wow_library/CMakeLists.txt
@@ -11,6 +11,8 @@ add_library(
 
 target_include_directories(${PROJECT_NAME} PUBLIC include ${CMAKE_CURRENT_SOURCE_DIR})
 
+target_link_libraries(${PROJECT_NAME} simulator)
+
 if (NOT EMSCRIPTEN)
     add_subdirectory(tests)
 endif ()
