+++ wow_library/source/Character.cpp
@@ -1,5 +1,7 @@
 #include "Character.hpp"
 
+#include "Helper_functions.hpp"
+
 #include <iostream>
 
 Character::Character(const Race& race, int level)
@@ -43,6 +45,48 @@ Character::Character(const Race& race, int level)
     default:
         assert(false);
     }
+}
+
+Character character_setup(const Armory& armory, const std::string& race, const std::vector<std::string>& armor_vec,
+                          const std::vector<std::string>& weapons_vec, const std::vector<std::string>& buffs_vec,
+                          const std::vector<std::string>& talent_string, const std::vector<int>& talent_val,
+                          const std::vector<std::string>& ench_vec)
+{
+    auto character = get_character_of_race(race);
+
+    character.equip_armor(armory.find_armor(Socket::head, armor_vec[0]));
+    character.equip_armor(armory.find_armor(Socket::neck, armor_vec[1]));
+    character.equip_armor(armory.find_armor(Socket::shoulder, armor_vec[2]));
+    character.equip_armor(armory.find_armor(Socket::back, armor_vec[3]));
+    character.equip_armor(armory.find_armor(Socket::chest, armor_vec[4]));
+    character.equip_armor(armory.find_armor(Socket::wrist, armor_vec[5]));
+    character.equip_armor(armory.find_armor(Socket::hands, armor_vec[6]));
+    character.equip_armor(armory.find_armor(Socket::belt, armor_vec[7]));
+    character.equip_armor(armory.find_armor(Socket::legs, armor_vec[8]));
+    character.equip_armor(armory.find_armor(Socket::boots, armor_vec[9]));
+    character.equip_armor(armory.find_armor(Socket::ring, armor_vec[10]));
+    character.equip_armor(armory.find_armor(Socket::ring, armor_vec[11]));
+    character.equip_armor(armory.find_armor(Socket::trinket, armor_vec[12]));
+    character.equip_armor(armory.find_armor(Socket::trinket, armor_vec[13]));
+    character.equip_armor(armory.find_armor(Socket::ranged, armor_vec[14]));
+
+    if (weapons_vec.size() > 1)
+    {
+        character.equip_weapon(armory.find_weapon(Weapon_socket::one_hand, weapons_vec[0]),
+                               armory.find_weapon(Weapon_socket::one_hand, weapons_vec[1]));
+    }
+    else
+    {
+        character.equip_weapon(armory.find_weapon(Weapon_socket::two_hand, weapons_vec[0]));
+    }
+
+    armory.add_enchants_to_character(character, ench_vec);
+    armory.add_buffs_to_character(character, buffs_vec);
+    armory.add_talents_to_character(character, talent_string, talent_val);
+
+    armory.compute_total_stats(character);
+
+    return character;
 }
 
 std::ostream& operator<<(std::ostream& os, const Character& character)
